requirements.txt
fastapi
requests
pytest-mock
uvicorn